LayoutTests/inspector/styles/media-emulation-expected.txt

Tests that emulated CSS media is reflected in the Styles pane.

Original style:
[expanded] 

[expanded] 
element.style  { ()

======== Matched CSS Rules ========
[expanded] 
#main  { (media-emulation.html:5)
color: red;

[expanded] 
div  { (user agent stylesheet)
display: block;


print media emulated:
[expanded] 

[expanded] 
element.style  { ()

======== Matched CSS Rules ========
[expanded] 
@media print (media-emulation.html:7)
#main  { (media-emulation.html:8)
color: black;

[expanded] 
#main  { (media-emulation.html:5)
/-- overloaded --/ color: red;

[expanded] 
div  { (user agent stylesheet)
display: block;


tty media emulated:
[expanded] 

[expanded] 
element.style  { ()

======== Matched CSS Rules ========
[expanded] 
@media tty (media-emulation.html:11)
#main  { (media-emulation.html:12)
color: green;

[expanded] 
#main  { (media-emulation.html:5)
/-- overloaded --/ color: red;

[expanded] 
div  { (user agent stylesheet)
display: block;


No media emulated:
[expanded] 

[expanded] 
element.style  { ()

======== Matched CSS Rules ========
[expanded] 
#main  { (media-emulation.html:5)
color: red;

[expanded] 
div  { (user agent stylesheet)
display: block;
